requirements.txt
albumentations==1.4.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
efficientnet_pytorch==0.7.1
filelock==3.13.1
fsspec==2024.3.0
gdown==5.1.0
huggingface-hub==0.22.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
joblib==1.3.2
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.2.0
pretrainedmodels==0.7.4
PySocks==1.7.1
PyYAML==6.0.1
rawpy==0.19.1
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
segmentation-models-pytorch==0.3.3
six==1.16.0
soupsieve==2.5
sympy==1.12
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.2
torch==2.2.1
torchsummary==1.5.1
torchvision==0.17.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
